tests/wpt/css-tests/css-exclusions-1_dev/xhtml1/wrap-flow-001.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
        <title>CSS Exclusions Test: wrap-flow property set to clear</title>
        <link href="mailto:jacobg@adobe.com" rel="author" title="Jacob Goldstein" />
        <link href="http://www.w3.org/TR/css3-exclusions/#wrap-flow-property" rel="help" />
        <meta content="ahem dom" name="flags" />
        <meta content="The text should flow above and below the exclusion, and not on either side" name="assert" />
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="resources/helper.js"></script>
        <style>
            * {
                margin: 0px;
                padding: 0px;
            }
            .exclusion {
                wrap-flow: clear;
                position: absolute;
                top: 40px;
                left: 50px;
                width: 110px;
                height: 110px;
                background: blue;
            }
            #content {
                width: 400px;
                height: 200px;
                line-height: 12px;
                font: 12px Ahem;
            }
        </style>
    </head>
    <body>
        <div class="exclusion"></div>

        <div id="content"><span id="linesAbove">A<br />B<br />C<br /></span><span id="linesBelow">D<br />E</span>
        </div>

        <div id="log"></div>

        <script type="text/javascript">
            test(function() {assert_equals(checkLinePos("linesBelow",150,"top"), true)}, "Verify top of the 'linesBelow' span is positioned correctly");
        </script>
    

</body></html>
